Extract MakerClient to own file
